Onyx Re-Write
It became clear that making a node for every base object wasn't particularly useful, so the entire Onyx node codebase was rewritten to be easier to read, understand and debug, and to allow multiple shapes to be generated with one node.

- All shapes have been replaced with OnyxShape.
- Currently only supports Box for testing purposes, all other shapes will be re-added quickly.
- Every shape generator sub-classes the OnyxGenerator class, and is responsible for control points, mesh generation and shape properties.
- The OnyxShape class saves required properties between scene loads.

The only remaining bug so far is an editor crash when a scene is closed with the object inside.